include: Move consumed selection recording
Include line replacement records consumed selections so later hunk
navigation can mask already handled replacement lines. The metadata store
also provides file-scoped reads and writes for that session state.

The recording workflow combines batch ownership translation, source
advancement, and command replacement context. Keeping that workflow in the
data store makes a persistence module import batch logic.

This commit addresses that by moving record_consumed_selection into
commands.selection.consumed_selection_recording and leaving
data.consumed_selections to read and write metadata records.
include_line_replacement imports the moved recorder from command support.

Subsequent commits will move the recording tests under the command
selection test tree and add architecture coverage for the new boundary.
